// dtos/register-data.dto.ts
export interface RegisterDataDTO {
  salutation: string;
  dateOfBirth: string;
  firstName: string;
  lastName: string;
  countryCode: string;
  phoneNumber: string;
  email: string;
  nationality: string;
  phoneNumberAddress: string;
  country: string;
  state: string;
  city: string;
  postalCode: string;
}